hahahah

+++ EventIF/src/br/edu/ifg/tads/mtp/eventif/DAO/EventoDao.java
@@ -90,7 +90,7 @@ public class EventoDao {
 			con = new ConnectionFactory().getConnection();
 			stmt = con.prepareStatement(sql, java.sql.Statement.RETURN_GENERATED_KEYS);
 
-			stmt.setInt(1, evento.getContato().getTelefone());
+			stmt.setLong(1, evento.getContato().getTelefone());
 			stmt.setString(2, evento.getContato().getEmail());
 
 			// executa
@@ -344,7 +344,7 @@ public class EventoDao {
 		try {
 			PreparedStatement stmt = new ConnectionFactory().getConnection()
 					.prepareStatement(sql);
-			stmt.setInt(1, cont.getTelefone());
+			stmt.setLong(1, cont.getTelefone());
 			stmt.setString(2, cont.getEmail());
 			stmt.executeUpdate();
 			stmt.close();
